Refactor single cell math
